Handle push to GistDetail instead of present modal
